Update fast-eloquent Laravel project from version 6 to 7
We do this because all new projects of this year must be compatible
with Laravel 8. So, we first update the project to Laravel 7 and
next to Laravel 8.

The steps are the following:

Make sure that PHP version is up to 7.2.5

Update dependencies in our composer.json file:

 * laravel/framework to ^7.0
 * nunomaduro/collision to ^4.1
 * phpunit/phpunit to ^8.5
 * laravel/tinker to ^2.0
 * facade/ignition to ^2.0

The following first-party package have new major releases to support
Laravel 7.

 * Telescope v3.0

Re-publish the assets of Telescope

Update dependencies in composer.json file again:

 * laravel/telescope to ^3.0

So, make sure that App\Excepton\Handler class accept instances of the
Throwable interface.

Update our session configuration file's option to have a fallback
value of null:

 'secure' => env('SESSION_SECURE_COOKIE', null),

Execute composer update.