Upgrade Laravel fast-eloquent project to 8.0 from 7.x
All new projects must be done in Laravel 8

The steps are the following:

Make sure, the new minimum version of PHP is 7.3.0

Updating dependencies that be used in our project:
 * laravel/framework to 8.0
 * nunomaduro/collision to 5.0
 * phpunit/phpunit to 9.0

Also, update first-party packages that be used in our project:

 * laravel/telescope to 4.0

Re-publish assets for Telescope, so that it works.

For running the seeders accordingly, we must create psr4 namespaces
and referencing it at the top all of the seeders classes.

For running the factories accordingly, require of
laravel/legacy-factories. So we  don't change factories code.

Keep using boostratp styles for pagination instead of Tailwind.

Allow controllers route declarations to use the standard PHP
callable syntax which provides better support for jumping to the
controller class in many IDEs.